+++ packages/lambda-tiler/src/tile.set.cache.ts
@@ -6,7 +6,13 @@ import { TileSetRaster } from './tile.set.raster';
 import { TileSetVector } from './tile.set.vector';
 
 export class TileSetCache {
-    cache: Map<string, Promise<TileSet | null>> = new Map();
+    /** Duration to cache tile sets for */
+    CacheTime = 30_000;
+
+    /** Cache the fetch requests */
+    cache: Map<string, { time: number; value: Promise<TileSet | null> }> = new Map();
+    /** Cache initialized tile sets */
+    tileSets: Map<string, TileSet> = new Map();
 
     id(tileSet: TileSet): string;
     id(name: string, tileMatrix: TileMatrixSet): string;
@@ -22,17 +28,19 @@ export class TileSetCache {
     add(tileSet: TileSet): void {
         const id = this.id(tileSet);
         if (this.cache.has(id)) throw new Error('Trying to add duplicate tile set:' + id);
-        this.cache.set(id, Promise.resolve(tileSet));
+        this.cache.set(id, { time: Date.now(), value: Promise.resolve(tileSet) });
     }
 
     get(name: string, tileMatrix: TileMatrixSet): Promise<TileSet | null> {
         const tsId = this.id(name, tileMatrix);
         let existing = this.cache.get(tsId);
-        if (existing == null) {
-            existing = this.loadTileSet(name, tileMatrix);
+        // Validate the data is current every ~30 seconds
+        if (existing == null || Date.now() - existing.time > this.CacheTime) {
+            const value = this.loadTileSet(name, tileMatrix);
+            existing = { time: Date.now(), value };
             this.cache.set(tsId, existing);
         }
-        return existing;
+        return existing.value;
     }
 
     async loadTileSet(name: string, tileMatrix: TileMatrixSet): Promise<TileSet | null> {
@@ -53,13 +61,22 @@ export class TileSetCache {
             return null;
         }
 
-        if (Config.TileSet.isRaster(tileSet)) {
+        // If we already have a copy and it hasn't been modified just return it
+        const existing = this.tileSets.get(tileSetId);
+        if (existing?.tileSet.updatedAt === tileSet.updatedAt) {
+            return existing;
+        }
+
+        if (Config.isTileSetRaster(tileSet)) {
             const ts = new TileSetRaster(name, tileMatrix);
             await ts.init(tileSet);
+            this.tileSets.set(tileSetId, ts);
             return ts;
         }
+
         const ts = new TileSetVector(name, tileMatrix);
         ts.tileSet = tileSet;
+        this.tileSets.set(tileSetId, ts);
         return ts;
     }
 
@@ -69,8 +86,8 @@ export class TileSetCache {
 
         const promises: Promise<TileSet | null>[] = [];
         for (const tileMatrix of tileMatrices) promises.push(this.get(name, tileMatrix));
-
         const tileMatrixSets = await Promise.all(promises);
+
         const tileSets: TileSetRaster[] = [];
         for (const parent of tileMatrixSets) {
             if (parent == null) continue;